requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.9.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
ruff==0.5.7
scipy==1.14.0
setuptools==73.0.1
six==1.16.0
trimesh==4.4.3
